test(gateway): split node provisioning coverage
